Add Current Weather

+++ app/src/main/java/com/paasta/newernews/domain/model/request/RequestGetCurrentWeather.kt
@@ -3,8 +3,8 @@ package com.paasta.newernews.domain.model.request
 import com.paasta.newernews.data.network.nnapi.BaseUrl
 
 data class RequestGetCurrentWeather(
-    var lat: Long,
-    var lon: Long,
+    var lat: Double,
+    var lon: Double,
     var appId: String = BaseUrl.WEATHER_API_KEY,
     var units: String = "metric"
 )
